tweaked the sequential root_agent workflow
Now, the 'root_agent' doesnt executes all of its sub_agents in a sequence.

Instead, a 'blog_pipeline' sequential agent is created and passed as tool to the root_agent.

If the user wishes to create a blog, then the root agent calls this pipeline.

If the user asks irrelevant questions (out-of-context), then the pipeline wont be called. thus the execution stops at the root_agent level itself.